Extract function and add tests
